Making good start. Cleaning up a lot of the code. About to start on Inheritance for the HealthSystem this will make it so much cleaner
