+++ SpringsValley/Assets/Scripts/Game/UIManager.cs
@@ -32,7 +32,7 @@ public class UIManager : MonoBehaviour
             SceneManager.LoadScene(SceneManager.GetActiveScene().buildIndex);
         }
 
-        if (HealthSystem.isDead && checkAlive)
+        if (checkAlive)
         {
             Debug.Log("Player is dead");
             checkAlive = false;
