+++ SpringsValley/Assets/Scripts/Player/PlayerControls.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 049da1ec00b50dd45943dd9917441886
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
