+++ SpringsValley/Assets/Scripts/Player/PlayerHealth.cs
@@ -1,0 +1,11 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class PlayerHealth : HealthSystem
+{
+    public UIManager uiManager;
+
+ 
+
+}
